backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
click==8.3.1
et_xmlfile==2.0.0
fastapi==0.122.0
h11==0.16.0
idna==3.11
mysql-connector-python==9.5.0
numpy==2.3.5
openpyxl==3.1.5
pandas==2.3.3
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
